+++ src/uversion.cpp
@@ -27,10 +27,7 @@ void init_uversion(py::module &m) {
 
   m.def(
       "u_version_to_string",
-      [](const std::vector<uint8_t> &version_array) {
-        if (version_array.size() != sizeof(UVersionInfo)) {
-          throw icupy::ICUError(U_ILLEGAL_ARGUMENT_ERROR);
-        }
+      [](const std::array<uint8_t, sizeof(UVersionInfo)> &version_array) {
         UVersionInfo info;
         std::copy(version_array.begin(), version_array.end(), info);
         char buf[U_MAX_VERSION_STRING_LENGTH];